Ajustes
- Migrando use enums pro core
- Ajustes nos logs de erro
- Adicionando readme ao projeto
