+++ marvel_heroes_core/lib/src/core/core.dart
@@ -1,5 +1,4 @@
 export '../core/libs/lib_dependencies.dart';
 export '../core/bindings/initial_bindings.dart';
 export '../core/loger/logger.dart';
-export '../core/method_channel/call_method_channel_enum.dart';
 export '../core/method_channel/method_channel_listener.dart';
