+++ marvel_heroes_core/lib/src/core/method_channel/method_channel_listener.dart
@@ -37,7 +37,7 @@ class MethodChannelListener {
         }
       });
     } catch (e) {
-      debugPrint(e.toString());
+      Log.error(e.toString());
     }
   }
 }
