+++ marvel_heroes_core/lib/src/data/data.dart
@@ -1,1 +1,3 @@
-
+export '../data/enums/call_method_channel_enum.dart';
+export '../data/enums/detail_type_enum.dart';
+export '../data/enums/order_by_enum.dart';
