+++ marvel_heroes_core/lib/src/data/enums/call_method_channel_enum.dart
