+++ marvel_heroes_core/lib/src/data/enums/detail_type_enum.dart
@@ -1,0 +1,9 @@
+enum DetailTypeEnum {
+  events('events'),
+  comics('comics'),
+  series('series');
+
+  final String value;
+
+  const DetailTypeEnum(this.value);
+}
