+++ marvel_heroes_core/lib/src/data/enums/http_error_type_enum.dart
@@ -1,0 +1,4 @@
+enum HttpErrorTypeEnum {
+  defaultError,
+  timeout,
+}
