+++ marvel_heroes_core/lib/src/data/enums/order_by_enum.dart
@@ -1,0 +1,9 @@
+enum OrderByEnum {
+  nameDesc('-name', 'Nome Z-A'),
+  nameAsc('name', 'Nome A-Z');
+
+  final String value;
+  final String desc;
+
+  const OrderByEnum(this.value, this.desc);
+}
