+++ marvel_heroes_home/lib/src/data/use_cases/get_hero_detail_by_type_use_case.dart
@@ -1,7 +1,7 @@
+import 'package:marvel_heroes_core/marvel_heroes_core.dart';
 import 'package:marvel_heroes_home/src/domain/entities/heroes_response_entity.dart';
 import 'package:marvel_heroes_home/src/domain/repositories/get_hero_detail_by_type_repository.dart';
 import 'package:marvel_heroes_home/src/domain/use_cases/get_hero_detail_by_type_use_case.dart';
-import 'package:marvel_heroes_home/src/infra/data_sources/home_datasource.dart';
 
 class GetHeroeDetailByTypeUsecase implements IGetHeroeDetailByTypeUsecase {
   final IGetHeroeDetailByTypeRepository _getHeroeDetailByTypeRepository;
