+++ marvel_heroes_home/lib/src/domain/repositories/get_hero_detail_by_type_repository.dart
@@ -1,5 +1,5 @@
+import 'package:marvel_heroes_core/marvel_heroes_core.dart';
 import 'package:marvel_heroes_home/src/domain/entities/heroes_response_entity.dart';
-import 'package:marvel_heroes_home/src/infra/data_sources/home_datasource.dart';
 
 abstract class IGetHeroeDetailByTypeRepository {
   Future<HeroesResponseEntity> getHeroDetailByType({
