+++ marvel_heroes_home/lib/src/domain/repositories/get_heroes_list_repository.dart
@@ -1,5 +1,5 @@
+import 'package:marvel_heroes_core/marvel_heroes_core.dart';
 import 'package:marvel_heroes_home/src/domain/entities/heroes_response_entity.dart';
-import 'package:marvel_heroes_home/src/presentation/pages/home/home_controller.dart';
 
 abstract class IGetHeroesListRepository {
   Future<HeroesResponseEntity> getHeroesList(
