+++ marvel_heroes_home/lib/src/infra/data_sources/home_datasource.dart
@@ -1,17 +1,6 @@
 import 'package:marvel_heroes_home/src/data/data_sources/home_datasource.dart';
 import 'package:marvel_heroes_core/marvel_heroes_core.dart';
 import 'package:marvel_heroes_home/src/data/models/heroes_response_model.dart';
-import 'package:marvel_heroes_home/src/presentation/pages/home/home_controller.dart';
-
-enum DetailTypeEnum {
-  events('events'),
-  comics('comics'),
-  series('series');
-
-  final String value;
-
-  const DetailTypeEnum(this.value);
-}
 
 class HomeDataSource implements IHomeDataSource {
   final IDioProvider _httpProvider;
