+++ marvel_heroes_home/test/src/domain/use_cases/get_hero_detail_by_type_use_case_test.dart
@@ -5,7 +5,6 @@ import 'package:marvel_heroes_home/src/data/repositories/get_hero_detail_by_type
 import 'package:marvel_heroes_home/src/data/use_cases/get_hero_detail_by_type_use_case.dart';
 import 'package:marvel_heroes_home/src/domain/repositories/get_hero_detail_by_type_repository.dart';
 import 'package:marvel_heroes_home/src/domain/use_cases/get_hero_detail_by_type_use_case.dart';
-import 'package:marvel_heroes_home/src/infra/data_sources/home_datasource.dart';
 
 import '../../mocks/home_data_source_mock.dart';
