+++ marvel_heroes_home/test/src/domain/use_cases/get_heroes_list_use_case_test.dart
@@ -6,8 +6,6 @@ import 'package:marvel_heroes_home/src/data/use_cases/get_heroes_list_use_case.d
 import 'package:marvel_heroes_home/src/domain/entities/heroes_response_entity.dart';
 import 'package:marvel_heroes_home/src/domain/repositories/get_heroes_list_repository.dart';
 import 'package:marvel_heroes_home/src/domain/use_cases/get_heroes_list_use_case.dart';
-import 'package:marvel_heroes_home/src/presentation/pages/home/home_controller.dart';
-
 import '../../mocks/home_data_source_mock.dart';
 
 void main() {
